+++ codeHF/o2_batch.sh
@@ -12,12 +12,14 @@ Index=0
 JSONLocal=$(basename $JSON)
 
 rm -f $FilesToMerge
+DirOutMain="output_o2"
+echo "Output directory: $DirOutMain (logfiles: $LogFile)"
 while read FileIn; do
   if [ ! -f "$FileIn" ]; then
-    echo "Error: Fle $FileIn does not exist."
+    echo "Error: File $FileIn does not exist."
     exit 1
   fi
-  DirOut="output_o2/$Index"
+  DirOut="$DirOutMain/$Index"
   mkdir -p $DirOut
   cd $DirOut
   cp "$JSON" $JSONLocal
@@ -27,21 +29,31 @@ while read FileIn; do
   echo "Input file ($Index): $FileIn"
   FileOut="$DirOut/$FILEOUT"
   echo $FileOut >> $DirBase/$FilesToMerge
-  echo "Output file: $FileOut"
+  #echo "Output file: $FileOut"
   bash $SCRIPT > $LogFile 2>&1 &
-  Index=$((Index+1))
+  PIDS+=($!)
+  ((Index+=1))
   cd $DirBase
 done < "$LISTINPUT"
-CmdNRun="top -u $USER -n 1 -c | grep o2 | grep configuration | wc -l"
-echo "Waiting for O2 to start..."
-while [ $(eval $CmdNRun) -eq 0 ]; do continue; done
+
 echo "Waiting for O2 to finish..."
-while [ $(eval $CmdNRun) -gt 0 ]; do
-  echo $(eval $CmdNRun)
-  sleep 1
+for pid in ${PIDS[@]}; do
+  wait $pid
+  STATUS+=($?)
 done
-sleep 1
-echo "Merging output files... (output file: $FILEOUT)"
+
+i=0
+ok=1
+for st in ${STATUS[@]}; do
+  if [ $st -ne 0 ]; then
+    echo "$i failed"
+    ok=0
+  fi
+  ((i+=1))
+done
+if [ $ok -ne 1 ]; then exit 1; fi # Exit if error.
+
+echo "Merging output files... (output file: $FILEOUT, logfile: $LogFile)"
 hadd $FILEOUT @"$FilesToMerge" > $LogFile 2>&1
 if [ $? -ne 0 ]; then echo "Error"; exit 1; fi # Exit if error.
 rm -f $FilesToMerge
